requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
Django==3.0.1
django-appconf==1.0.3
django-compressor==2.3
django-crispy-forms==1.8.1
django-extensions==2.2.5
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
Pillow==6.2.1
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
